+++ examples/fastapi_full/app/deps.py
@@ -7,6 +7,7 @@ __all__ = [
     "get_current_api_key_writer",
     "get_current_active_reader",
     "get_current_active_writer",
+    "get_current_active_admin",
     "get_current_active_superuser",
 ]
 
@@ -17,5 +18,6 @@ from fastapi_full.core.deps import (
     get_current_api_key_writer as get_current_api_key_writer,
     get_current_active_reader as get_current_active_reader,
     get_current_active_writer as get_current_active_writer,
+    get_current_active_admin as get_current_active_admin,
     get_current_active_superuser as get_current_active_superuser,
 )